requirements.txt
django==5.0
pymongo==4.10.1
djangorestframework==3.15.2
django-environ==0.11.2
requests==2.32.3
